+++ app/features/idl/interactive-idl/model/transaction/index.ts
@@ -1,5 +1,5 @@
 export { buildTransaction } from './build-transaction';
 export { formatTransactionError } from './format-transaction-error';
-export type { InstructionInvocationResult, SimulationResult } from './types';
+export type { InstructionInvocationResult, InstructionSimulationResult } from './types';
 export { useInvokeTransaction } from './use-invoke-transaction';
 export { useSimulateTransaction } from './use-simulate-transaction';
