+++ app/features/idl/interactive-idl/ui/InstructionActivity.tsx
@@ -1,26 +1,49 @@
 import { useExplorerLink } from '@entities/cluster';
 import { ProgramLogs, TxErrorStatus, TxSuccessStatus } from '@entities/program-logs';
+import { Badge } from '@shared/ui/badge';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@shared/ui/tabs';
 import { ReactNode } from 'react';
 
+import { Copyable } from '@/app/components/common/Copyable';
 import { Card } from '@/app/shared/ui/Card';
 import type { InstructionLogs } from '@/app/utils/program-logs';
 
-import type { InstructionInvocationResult } from '../model/transaction/types';
+import type { InstructionInvocationResult, InstructionSimulationResult } from '../model/transaction/types';
 
-type InstructionActivityProps = {
+type InstructionInvocationActivityProps = {
     lastResult?: InstructionInvocationResult;
-    logs: string[];
     parseLogs: (logs: string[]) => InstructionLogs[];
 };
 // FIXME: missing Storybook story — pure props, but uses useExplorerLink internally so needs withCluster decorator.
-export function InstructionActivity({ lastResult, logs, parseLogs }: InstructionActivityProps) {
+export function InstructionInvocationActivity({ lastResult, parseLogs }: InstructionInvocationActivityProps) {
     const tabs = [
         {
             component: (
                 <ProgramLogs
                     header={lastResult && <TxStatusHeader lastResult={lastResult} />}
-                    logs={logs}
+                    logs={lastResult?.logs ?? []}
+                    parseLogs={parseLogs}
+                />
+            ),
+            id: 'program-logs',
+            title: 'Program logs',
+        },
+    ];
+    return <CardWithTabs tabs={tabs} />;
+}
+
+type InstructionSimulationActivityProps = {
+    lastSimulation?: InstructionSimulationResult;
+    parseLogs: (logs: string[]) => InstructionLogs[];
+};
+
+export function InstructionSimulationActivity({ lastSimulation, parseLogs }: InstructionSimulationActivityProps) {
+    const tabs = [
+        {
+            component: (
+                <ProgramLogs
+                    header={lastSimulation && <SimulationStatusHeader lastSimulation={lastSimulation} />}
+                    logs={lastSimulation?.logs ?? []}
                     parseLogs={parseLogs}
                 />
             ),
@@ -72,5 +95,68 @@ function TxStatusHeader({ lastResult }: { lastResult: NonNullable<InstructionInv
             date={lastResult.finishedAt}
             link={lastResult.serializedTxMessage ? link : null}
         />
+    );
+}
+
+function SimulationStatusHeader({ lastSimulation }: { lastSimulation: NonNullable<InstructionSimulationResult> }) {
+    return lastSimulation.status === 'success' ? (
+        <SimulationSuccessStatus unitsConsumed={lastSimulation.unitsConsumed} date={lastSimulation.finishedAt} />
+    ) : (
+        <SimulationErrorStatus message={lastSimulation.message} date={lastSimulation.finishedAt} />
+    );
+}
+
+function formatTimestamp(date: Date): string {
+    const time = date.toLocaleTimeString('en-US', {
+        hour: '2-digit',
+        hour12: false,
+        minute: '2-digit',
+        second: '2-digit',
+        timeZone: 'UTC',
+    });
+    return `${time} UTC`;
+}
+
+function SimulationSuccessStatus({ unitsConsumed, date }: { unitsConsumed: number | undefined; date: Date }) {
+    return (
+        <div className="e-flex e-items-center e-gap-2 e-rounded e-border e-border-solid e-border-neutral-600 e-px-4 e-py-2">
+            {unitsConsumed !== undefined && (
+                <div className="e-flex e-items-center e-gap-1">
+                    <span className="e-text-xs e-tracking-tight e-text-accent-700">
+                        {unitsConsumed.toLocaleString('en-US')} CU
+                    </span>
+                </div>
+            )}
+            <div className="e-flex e-items-center">
+                <span className="e-whitespace-nowrap e-text-xs e-tracking-tight e-text-accent-700">
+                    {formatTimestamp(date)}
+                </span>
+            </div>
+            <Badge variant="success" size="xs" className="e-ml-auto">
+                Simulated
+            </Badge>
+        </div>
+    );
+}
+
+function SimulationErrorStatus({ message, date }: { message: string; date: Date }) {
+    return (
+        <div className="e-flex e-items-center e-gap-2 e-rounded e-border e-border-solid e-border-neutral-600 e-px-4 e-py-2">
+            <div className="e-flex e-w-1/2 e-items-center e-gap-1">
+                <Copyable text={message}>
+                    <span className="e-overflow-hidden e-text-ellipsis e-whitespace-nowrap e-font-mono e-text-sm e-tracking-tight e-text-destructive">
+                        {message}
+                    </span>
+                </Copyable>
+            </div>
+            <div className="e-flex e-items-center">
+                <span className="e-whitespace-nowrap e-text-xs e-tracking-tight e-text-destructive">
+                    {formatTimestamp(date)}
+                </span>
+            </div>
+            <Badge variant="destructive" size="xs" className="e-ml-auto">
+                Simulation Error
+            </Badge>
+        </div>
     );
 }
